Cron: extract execution orchestrator and adapters
Move cron execution-path branching out of the timer service into explicit execution modules: shared execution types and errors, a single orchestrator, and dedicated main-session and detached-session adapters. Keep the public timer exports and existing runtime behavior intact so this commit is purely a structural seam for later stages.

Regeneration-Prompt: |
  Continue the staged cron and heartbeat refactor after durable run records are in place. The next step is to stop letting src/cron/service/timer.ts directly own the main-session versus detached execution branch logic. Extract the execution logic into explicit modules under src/cron/execution so later stages can change adapters without mixing scheduler code and execution code. Preserve the existing external behavior, keep timer exports compatible for current callers and tests, and keep the change mostly structural. Shared timeout/error normalization should live with the execution seam so both the orchestrator and timer-facing callers use the same semantics.
