+++ src/cron/execution/types.ts
@@ -1,0 +1,9 @@
+import type { CronRunOutcome, CronRunTelemetry } from "../types.js";
+
+export type CronExecutionResult = CronRunOutcome &
+  CronRunTelemetry & {
+    delivered?: boolean;
+    deliveryAttempted?: boolean;
+  };
+
+export type WaitWithAbort = (ms: number) => Promise<void>;
